requirements.txt
Flask
pillow
Werkzeug
